% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rng.R
\name{rng_state_long_jump}
\alias{rng_state_long_jump}
\title{Advance random number state}
\usage{
rng_state_long_jump(state, times = 1L)
}
\arguments{
\item{state}{A raw vector representing \code{xoshiro} random number
generator; see \code{\link{rng}}.}

\item{times}{An integer indicating the number of times the
\code{long_jump} should be performed. The default is one, but values
larger than one will repeatedly advance the state.}
}
\description{
Advance a saved random number state by performing a "long jump" on
it. Intended for if you have serialised an RNG state (using the
\verb{$state()} method, or via some application using \code{xoshiro}) but want
create a new seed that is uncorrelated.  If the state extracted is
to be reused multiple times, then the state needs jumping to
prevent generating the same sequence of random numbers.
}
\examples{
# Create a new RNG object
rng <- xoshiro::rng$new(1)

# Serialise the state as a raw vector
state <- rng$state()

# We can advance this state
xoshiro::rng_state_long_jump(state)

# Which gives the same result as long_jump on the original generator
rng$long_jump()$state()
rng$long_jump()$state()

# Multiple jumps can be taken by using the "times" argument
xoshiro::rng_state_long_jump(state, 2)
}
